requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.7.0
google-genai>=0.8.0
httpx>=0.27.0
python-dateutil>=2.9.0
python-dotenv>=1.0.0
